services/gender_translation/pyproject.toml
[tool.poetry]
name = "gender_translation"
version = "0.1.0"
description = "Gender translation microservice"
authors = ["Stankevich Andrey <stankevich.as@phystech.edu>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
numpy = "1.26.4"
torch = {version = "^2.1.0", source = "pytorch-gpu"}
python-telegram-bot = { extras = ["webhooks"], version = "^22.3" }
httpx = "^0.27.0"
pydantic = "^2.8"
mlflow_registry = { path = "../../mlflow_registry", develop = true }
vision = { path = "../../vision", develop = true }
models = { path = "../../models", develop = true }
pydantic-settings = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-cov = "^5.0"
mypy = "^1.10"
flake8 = "^7.0.0"
flake8-import-order = "^0.19.2"
flake8-docstrings = "^1.7.0"

# PyTorch GPU source
pytest-asyncio = "^1.2.0"
[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true
disallow_untyped_defs = true
exclude = '(^\.venv/|^build/|^dist/)'

[tool.pytest.ini_options]
testpaths = ["../../tests/services/gender_translation"]
python_files = ["test_*.py"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
